+++ tests/structs.rs
@@ -39,7 +39,7 @@ fn atomic_struct() {
     // Check that metadata is correct.
     let metadata = settings.metadata();
     assert_eq!(metadata.max_depth, 2);
-    assert_eq!(metadata.max_topic_size, "c".len());
+    assert_eq!(metadata.max_length, "c".len());
 }
 
 #[test]
@@ -77,7 +77,7 @@ fn recursive_struct() {
     // Check that metadata is correct.
     let metadata = settings.metadata();
     assert_eq!(metadata.max_depth, 3);
-    assert_eq!(metadata.max_topic_size, "c/a".len());
+    assert_eq!(metadata.max_length, "c/a".len());
 }
 
 #[test]