+++ CHANGELOG.md
@@ -14,9 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fields in `#[derive(Miniconf)]` are now atomic by default. To recurse, users must
   annotate fields with `#[miniconf(defer)]`
 * New `miniconf::Option` type has been added. Existing `Option` implementation has been changed to
-  allow run-time nullability of settings for more flexibility.
+  allow run-time nullability of values for more flexibility.
 * New `miniconf::Array` type has been added, replacing the previous [T; N] implementation
 * `Miniconf` implementation on most primitive types has been removed as it is no longer required.
+* [breaking] The API has changed to be agnostic to usage (e.g. now referring to namespace paths and values
+  instead of topics and settings). Functions in the `Miniconf` trait have been renamed.
 
 ### Fixed
 * Python device discovery now only discovers unique device identifiers. See [#97](https://github.com/quartiq/miniconf/issues/97)